requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
flask-mail==0.9.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
opencv-python==4.5.4.58
tqdm==4.62.3